requirements.txt
requests>=2.31
mysql-connector-python>=8.0
schedule>=1.2
python-dotenv>=1.0
